src/main/java/com/github/karixdev/polslcoursescheduleapi/auth/AuthService.java
package com.github.karixdev.polslcoursescheduleapi.auth;

import com.github.karixdev.polslcoursescheduleapi.auth.payload.request.RegisterRequest;
import com.github.karixdev.polslcoursescheduleapi.shared.payload.response.SuccessResponse;
import com.github.karixdev.polslcoursescheduleapi.user.UserRole;
import com.github.karixdev.polslcoursescheduleapi.user.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;

@Service
@RequiredArgsConstructor
public class AuthService {
    private final UserService userService;

    @Transactional
    public SuccessResponse register(RegisterRequest payload) {
        userService.createUser(
                payload.getEmail(),
                payload.getPassword(),
                UserRole.ROLE_USER,
                false
        );

        return new SuccessResponse();
    }
}
